✨ (Query): Ajout liste gestionnaire au readModel et typage
