+++ src/modules/authN/Permission.ts
@@ -9,7 +9,6 @@ import {
   PermissionRetirerGF,
   PermissionListerProjets,
   PermissionExporterProjets,
-  PermissionModifierIdentifiantGestionnaireReseau,
 } from '@modules/project';
 import { PermissionInviterDgecValidateur } from '@modules/utilisateur';
 import { PermissionListerDemandesAdmin } from '@modules/modificationRequest';
@@ -18,6 +17,7 @@ import {
   PermissionListerGestionnairesRéseau,
   PermissionConsulterGestionnaireRéseau,
 } from '@modules/gestionnaireRéseau';
+import { PermissionModifierIdentifiantGestionnaireReseau } from '@modules/project/queries/ModifierGestionnaireRéseau';
 
 export type Permission = {
   nom: string;
