+++ src/modules/gestionnaireRéseau/lister/ListerDétailGestionnairesRéseau.ts
@@ -1,0 +1,6 @@
+import { ConsulterGestionnaireRéseauReadModel } from '../consulter';
+
+export type ListeDétailGestionnairesRéseauReadModel = Array<ConsulterGestionnaireRéseauReadModel>;
+
+export type ListerDétailGestionnairesRéseauQueryHandler =
+  () => Promise<ListeDétailGestionnairesRéseauReadModel>;
