+++ src/modules/gestionnaireRéseau/lister/index.ts
@@ -1,1 +1,2 @@
+export * from './ListerDétailGestionnairesRéseau';
 export * from './ListerGestionnairesRéseau';
